+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add rename functionality to VcsList and CollectionComponent
 - Add property `customClasses` to all vcs buttons for custom styling using css classes
 - Add multi-view plugin to external plugins and demo application
+- Add position display to footer
 
 # 5.0.3